+++ backend/canisters/group_index/impl/src/updates/c2c_mark_group_import_complete.rs
@@ -27,6 +27,7 @@ fn c2c_mark_group_import_complete_impl(args: Args, state: &mut RuntimeState) -> 
         Some(CommunityImportedInto {
             community_name: args.community_name,
             community_id,
+            local_user_index_canister_id: args.local_user_index_canister_id,
             channel: args.channel,
             other_default_channels: args.other_public_channels,
         }),